requirements.txt
streamlit
PyPDF2
pdfplumber
python-docx
pytesseract
Pillow
spacy 